refactor: Move check_and_set_session() to ECU class
