+++ src/gallia/uds/ecu.py
@@ -11,9 +11,13 @@ from gallia.transports.base import BaseTransport
 from gallia.uds.core import service
 from gallia.uds.core.client import UDSClient, UDSRequestConfig
 from gallia.uds.core.constants import DataIdentifier
-from gallia.uds.core.exception import ResponseException
+from gallia.uds.core.exception import ResponseException, UnexpectedNegativeResponse
 from gallia.uds.core.utils import from_bytes
-from gallia.uds.helpers import as_exception, raise_for_error
+from gallia.uds.helpers import (
+    as_exception,
+    raise_for_error,
+    suggests_identifier_not_supported,
+)
 from gallia.utils import g_repr
 
 
@@ -112,6 +116,82 @@ class ECU(UDSClient):
         Returns:
             True on success, False on error.
         """
+
+    async def check_and_set_session(
+        self,
+        expected_session: int,
+        retries: int = 3,
+    ) -> bool:
+        """check_and_set_session() reads the current session and (re)tries to set
+        the session to the expected session if they do not match.
+
+        Returns True if the current session matches the expected session,
+        or if read_session is not supported by the ECU or in the current session."""
+
+        self.logger.log_debug(
+            f"Checking current session, expecting {g_repr(expected_session)}"
+        )
+
+        try:
+            current_session = await self.read_session(
+                config=UDSRequestConfig(max_retry=retries)
+            )
+        except UnexpectedNegativeResponse as e:
+            if suggests_identifier_not_supported(e.RESPONSE_CODE):
+                self.logger.log_info(
+                    f"Read current session not supported: {e.RESPONSE_CODE.name}, skipping check_session"
+                )
+                return True
+            raise e
+        except asyncio.TimeoutError:
+            self.logger.log_warning(
+                "Reading current session timed out, skipping check_session"
+            )
+            return True
+
+        self.logger.log_debug(f"Current session is {g_repr(current_session)}")
+        if current_session == expected_session:
+            return True
+
+        for i in range(retries):
+            self.logger.log_warning(
+                f"Not in session {g_repr(expected_session)}, ECU replied with {g_repr(current_session)}"
+            )
+
+            self.logger.log_info(
+                f"Switching to session {g_repr(expected_session)}; attempt {i + 1} of {retries}"
+            )
+            resp = await self.set_session(expected_session)
+
+            if isinstance(resp, service.NegativeResponse):
+                self.logger.log_warning(
+                    f"Switching to session {g_repr(expected_session)} failed: {resp}"
+                )
+
+            try:
+                current_session = await self.read_session(
+                    config=UDSRequestConfig(max_retry=retries)
+                )
+                self.logger.log_debug(f"Current session is {g_repr(current_session)}")
+                if current_session == expected_session:
+                    return True
+            except UnexpectedNegativeResponse as e:
+                if suggests_identifier_not_supported(e.RESPONSE_CODE):
+                    self.logger.log_info(
+                        f"Read current session not supported: {e.RESPONSE_CODE.name}, skipping check_session"
+                    )
+                    return True
+                raise e
+            except asyncio.TimeoutError:
+                self.logger.log_warning(
+                    "Reading current session timed out, skipping check_session"
+                )
+                return True
+
+        self.logger.log_warning(
+            f"Failed to switch to session {g_repr(expected_session)} after {retries} attempts"
+        )
+        return False
 
     async def power_cycle(self, sleep: int = 5) -> bool:
         if self.power_supply is None:
