+++ src/gallia/udscan/scanner/scan_services.py
@@ -10,7 +10,6 @@ from gallia.uds.core.exception import UDSException
 from gallia.uds.core.service import NegativeResponse, UDSResponse
 from gallia.uds.helpers import suggests_service_not_supported
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import check_and_set_session
 from gallia.utils import ParseSkips, auto_int, g_repr
 
 
@@ -127,7 +126,7 @@ class ScanServices(UDSScanner):
                     continue
 
                 if args.check_session:
-                    if not await check_and_set_session(self.ecu, session):
+                    if not await self.ecu.check_and_set_session(session):
                         self.logger.log_error(
                             f"Aborting scan on session {g_repr(session)}; current SID was {g_repr(sid)}"
                         )
