+++ src/gallia/udscan/scanner/simple_rtcl.py
@@ -6,7 +6,6 @@ from typing import Union
 
 from gallia.uds.core.service import NegativeResponse, RoutineControlResponse
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import check_and_set_session
 from gallia.utils import auto_int, g_repr
 
 
@@ -79,7 +78,7 @@ class RTCL(UDSScanner):
 
     async def main(self, args: Namespace) -> None:
         try:
-            await check_and_set_session(self.ecu, args.session)
+            await self.ecu.check_and_set_session(args.session)
         except Exception as e:
             self.logger.log_critical(
                 f"Could not change to session: {g_repr(args.session)}: {g_repr(e)}"
