+++ src/gallia/udscan/scanner/simple_wmba.py
@@ -5,7 +5,6 @@ from pathlib import Path
 
 from gallia.uds.core.service import NegativeResponse
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import check_and_set_session
 from gallia.utils import auto_int, g_repr
 
 
@@ -36,7 +35,7 @@ class WriteMemoryByAddressScanner(UDSScanner):
 
     async def main(self, args: Namespace) -> None:
         try:
-            await check_and_set_session(self.ecu, args.session)
+            await self.ecu.check_and_set_session(args.session)
         except Exception as e:
             self.logger.log_critical(
                 f"Could not change to session: {g_repr(args.session)}: {g_repr(e)}"
